Complete Rewrite of the Dictionary
Abandoned the original structure as it was getting far too messy. Emulating the behaviors and structure of the Python Dictionary, this is more realistic and likely more advanced than my original attempt at implementation. Known issues are that the nationalities do not appear to be alphabetized despite my code attempting to do so. However, the nationalities dictionary is getting closer to  completion.